cambio de funciones al nuevo archivo SubmissionFeedback
